Scheduling Assistant: Custom Special Registration
- if special registration is closed, but the student has some not-canceled override requests
  (/checkEligibility returns eligible:false and hasNonCanceledRequest:true)
  - show the Requested Overrides table
  - requests cannot be canceled by student at this point

+++ JavaSource/org/unitime/timetable/gwt/client/sectioning/SpecialRegistrationsPanel.java
@@ -279,7 +279,7 @@ public class SpecialRegistrationsPanel extends P {
 				}
 			}
 			ImageButton delete = null;
-			if (reg.canCancel()) {
+			if (reg.canCancel() && iSpecReg.isEnabled()) {
 				delete = new ImageButton(RESOURCES.delete(), RESOURCES.delete_Down(), RESOURCES.delete_Over());
 				delete.addStyleName("unitime-NoPrint");
 				delete.addClickHandler(new ClickHandler() {
